feat: batch build command
