+++ .gitignore
@@ -17,6 +17,7 @@
 
 # Local
 templar
+templar-*
 /templates/
 *.env
 *.txt
